+++ lib/components/base/BaseHeading.vue
@@ -14,6 +14,9 @@ component(
     }
   ]`
   :is="tag"
+  :style=`{
+    fontSize: computedFontSize
+  }`
 ): slot
 </template>
 
@@ -34,6 +37,10 @@ export default {
           ) !== -1
         );
       }
+    },
+    fontSize: {
+      type: String,
+      default: null
     },
     tag: {
       type: String,
@@ -78,6 +85,28 @@ export default {
       }
 
       return color;
+    },
+
+    computedFontSize() {
+      if (this.fontSize) {
+        return this.fontSize;
+      } else if (this.tag === "h1") {
+        return "26px";
+      } else if (this.tag === "h2") {
+        return "24px";
+      } else if (this.tag === "h3") {
+        return "22px";
+      } else if (this.tag === "h4") {
+        return "20px";
+      } else if (this.tag === "h5") {
+        return "18px";
+      } else if (this.tag === "h6") {
+        return "16px";
+      } else if (this.tag === "p") {
+        return "16px";
+      } else if (this.tag === "small") {
+        return "14px";
+      }
     },
 
     computedWeight() {
@@ -125,44 +154,36 @@ $colors: black, blue, green, grey, red, orange, white;
   // --> TAGS <--
 
   &--h1 {
-    font-size: 26px;
     line-height: 36px;
   }
 
   &--h2 {
-    font-size: 24px;
     line-height: 34px;
   }
 
   &--h3 {
-    font-size: 22px;
     line-height: 32px;
   }
 
   &--h4 {
-    font-size: 20px;
     line-height: 30px;
   }
 
   &--h5 {
-    font-size: 18px;
     line-height: 28px;
   }
 
   &--h6 {
-    font-size: 16px;
     line-height: 28px;
   }
 
   &--p {
     margin-bottom: 20px;
-    font-size: 16px;
     line-height: 28px;
   }
 
   &--small {
     margin-bottom: 20px;
-    font-size: 14px;
     line-height: 24px;
   }